majo project/security-incident-chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
openai==1.3.0
jira==3.5.1
slack_sdk==3.23.0
pydantic==2.4.2
jinja2==3.1.2 